1. Dodanie bicycle_id do tabeli orders, aby sprzedawca mógł jednoznacznie przypisać rower do zamówienia. 2. Inne zmiany porządkujące projekt, takie jak dodanie do data.sql polecenia, które po wywołaniu w microsoft SQL Management studio wczyta dane przykładowych 13 rowerów, na wypadek problemów z połączeniem bazą danych.

+++ src/main/java/pl/edu/wszib/jwd/onlinestore/web/models/BicycleModel.java
@@ -118,7 +118,7 @@ public class BicycleModel {
         return wheelSize;
     }
 
-    public void setWheelSize(int wheelSize) {
+    public void setWheelSize(float wheelSize) {
         this.wheelSize = wheelSize;
     }
